refactor logic and creation
