+++ ivy/functional/frontends/jax/numpy/__init__.py
@@ -404,10 +404,10 @@ from . import creation
 from .creation import *
 from . import dtype
 from .dtype import can_cast, promote_types
-from . import comparison
-from .comparison import *
 from . import general
 from .general import *
+from . import logic
+from .logic import *
 from . import manipulations
 from .manipulations import *
 from . import mathematical_functions
