+++ ivy/functional/frontends/jax/numpy/manipulations.py
@@ -100,17 +100,6 @@ def take(
     return ivy.take_along_axis(a, indices, axis, out=out)
 
 
-@handle_numpy_dtype
-@to_ivy_arrays_and_back
-def asarray(a, dtype=None, order=None):
-    return ivy.asarray(a, dtype=dtype)
-
-
 @to_ivy_arrays_and_back
 def broadcast_to(arr, shape):
     return ivy.broadcast_to(arr, shape)
-
-
-@to_ivy_arrays_and_back
-def uint16(x):
-    return ivy.astype(x, ivy.UintDtype("uint16"), copy=False)
