+++ ivy_tests/test_ivy/test_frontends/test_jax/test_jax_numpy_general.py
@@ -4,148 +4,7 @@ import numpy as np
 
 # local
 import ivy_tests.test_ivy.helpers as helpers
-import ivy_tests.test_ivy.test_frontends.test_numpy.helpers as np_helpers
 from ivy_tests.test_ivy.helpers import handle_frontend_test
-
-
-# all
-@handle_frontend_test(
-    fn_tree="jax.numpy.all",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=st.one_of(st.just(("bool",)), helpers.get_dtypes("integer")),
-    ),
-)
-def test_jax_numpy_all(
-    *,
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        all_aliases=["numpy.alltrue"],
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-    )
-
-
-# argmax
-@handle_frontend_test(
-    fn_tree="jax.numpy.argmax",
-    dtype_and_x=helpers.dtype_values_axis(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        force_int_axis=True,
-        min_num_dims=1,
-        valid_axis=True,
-    ),
-    keepdims=st.booleans(),
-)
-def test_jax_numpy_argmax(
-    *,
-    dtype_and_x,
-    keepdims,
-    as_variable,
-    with_out,
-    num_positional_args,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x, axis = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=with_out,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        axis=axis,
-        out=None,
-        keepdims=keepdims,
-    )
-
-
-# argwhere
-@handle_frontend_test(
-    fn_tree="jax.numpy.argwhere",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("valid"),
-    ),
-)
-def test_jax_numpy_argwhere(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-    on_device,
-):
-    dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        size=None,
-        fill_value=None,
-    )
-
-
-# argsort
-@handle_frontend_test(
-    fn_tree="jax.numpy.argsort",
-    dtype_x_axis=helpers.dtype_values_axis(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        min_axis=-1,
-        max_axis=0,
-        min_num_dims=1,
-        force_int_axis=True,
-    ),
-)
-def test_jax_numpy_argsort(
-    *,
-    dtype_x_axis,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-    on_device,
-):
-    input_dtype, x, axis = dtype_x_axis
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        axis=axis,
-    )
 
 
 # fmax
@@ -181,134 +40,6 @@ def test_jax_numpy_fmax(
         on_device=on_device,
         x1=inputs[0],
         x2=inputs[1],
-    )
-
-
-# bitwise_and
-# TODO: add testing for other dtypes
-@handle_frontend_test(
-    fn_tree="jax.numpy.bitwise_and",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("bool"), num_arrays=2
-    ),
-)
-def test_jax_numpy_bitwise_and(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# bitwise_not
-@handle_frontend_test(
-    fn_tree="jax.numpy.bitwise_not",
-    dtype_and_x=helpers.dtype_and_values(available_dtypes=helpers.get_dtypes("bool")),
-)
-def test_jax_numpy_bitwise_not(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x=x[0],
-    )
-
-
-# bitwise_or
-# TODO: add testing for other dtypes
-@handle_frontend_test(
-    fn_tree="jax.numpy.bitwise_or",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("bool"), num_arrays=2
-    ),
-)
-def test_jax_numpy_bitwise_or(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
-    )
-
-
-# bitwise_xor
-# TODO: add testing for other dtypes
-@handle_frontend_test(
-    fn_tree="jax.numpy.bitwise_xor",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("bool"), num_arrays=2
-    ),
-)
-def test_jax_numpy_bitwise_xor(
-    *,
-    dtype_and_x,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        x1=x[0],
-        x2=x[1],
     )
 
 
@@ -349,57 +80,6 @@ def test_jax_numpy_heaviside(
         on_device=on_device,
         x1=x[0],
         x2=x[0],
-    )
-
-
-# any
-@handle_frontend_test(
-    fn_tree="jax.numpy.any",
-    dtype_x_axis=helpers.dtype_values_axis(
-        available_dtypes=helpers.get_dtypes("valid", full=True),
-        valid_axis=True,
-        max_axes_size=1,
-        force_int_axis=True,
-    ),
-    keepdims=st.booleans(),
-    where=np_helpers.where(),
-)
-def test_jax_numpy_any(
-    *,
-    dtype_x_axis,
-    keepdims,
-    where,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    input_dtype, x, axis = dtype_x_axis
-    if isinstance(axis, tuple):
-        axis = axis[0]
-    where, as_variable, native_array = np_helpers.handle_where_and_array_bools(
-        where=where,
-        input_dtype=input_dtype,
-        as_variable=as_variable,
-        native_array=native_array,
-    )
-    np_helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        all_aliases=["numpy.sometrue"],
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        axis=axis,
-        out=None,
-        keepdims=keepdims,
-        where=where,
     )
 
 
@@ -646,228 +326,4 @@ def test_jax_numpy_diag(
         on_device=on_device,
         v=x[0],
         k=k,
-    )
-
-
-# zeros_like
-@handle_frontend_test(
-    fn_tree="jax.numpy.zeros_like",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("valid"),
-    ),
-    shape=helpers.get_shape(
-        allow_none=True,
-        min_num_dims=1,
-        max_num_dims=5,
-        min_dim_size=1,
-        max_dim_size=10,
-    ),
-    dtype=helpers.get_dtypes("valid", full=False),
-)
-def test_numpy_zeros_like(
-    dtype_and_x,
-    dtype,
-    shape,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-    on_device,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        dtype=dtype[0],
-        shape=shape,
-    )
-
-
-# msort
-@handle_frontend_test(
-    fn_tree="jax.numpy.msort",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("numeric"),
-        min_num_dims=2,
-        min_dim_size=2,
-    ),
-)
-def test_jax_numpy_msort(
-    dtype_and_x,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        a=x[0],
-    )
-
-
-# arange
-@handle_frontend_test(
-    fn_tree="jax.numpy.arange",
-    start=st.integers(min_value=-100, max_value=100),
-    stop=st.integers(min_value=-100, max_value=100) | st.none(),
-    step=st.integers(min_value=-100, max_value=100).filter(lambda x: x != 0),
-    dtype=helpers.get_dtypes("numeric", full=False),
-)
-def test_jax_numpy_arange(
-    *,
-    start,
-    stop,
-    step,
-    dtype,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    helpers.test_frontend_function(
-        input_dtypes=dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        start=start,
-        stop=stop,
-        step=step,
-        dtype=dtype[0],
-    )
-
-
-# zeros
-@handle_frontend_test(
-    fn_tree="jax.numpy.zeros",
-    shape=helpers.get_shape(
-        allow_none=False,
-        min_num_dims=1,
-        max_num_dims=5,
-        min_dim_size=1,
-        max_dim_size=10,
-    ),
-    dtypes=helpers.get_dtypes("numeric", full=False),
-)
-def test_jax_numpy_zeros(
-    *,
-    dtypes,
-    shape,
-    num_positional_args,
-    as_variable,
-    native_array,
-    on_device,
-    fn_tree,
-    frontend,
-):
-    helpers.test_frontend_function(
-        input_dtypes=dtypes,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        shape=shape,
-        dtype=dtypes[0],
-    )
-
-
-# ones
-@handle_frontend_test(
-    fn_tree="jax.numpy.ones",
-    shape=helpers.get_shape(
-        allow_none=False,
-        min_num_dims=1,
-        max_num_dims=5,
-        min_dim_size=1,
-        max_dim_size=10,
-    ),
-    dtype=helpers.get_dtypes("valid", full=False),
-)
-def test_jax_numpy_ones(
-    shape,
-    dtype,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-    on_device,
-):
-    helpers.test_frontend_function(
-        input_dtypes=dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        shape=shape,
-        dtype=dtype[0],
-    )
-
-
-# ones_like
-@handle_frontend_test(
-    fn_tree="jax.numpy.ones_like",
-    dtype_and_x=helpers.dtype_and_values(
-        available_dtypes=helpers.get_dtypes("valid"),
-    ),
-    shape=helpers.get_shape(
-        allow_none=True,
-        min_num_dims=1,
-        max_num_dims=5,
-        min_dim_size=1,
-        max_dim_size=10,
-    ),
-    dtype=helpers.get_dtypes("valid", full=False),
-)
-def test_jax_numpy_ones_like(
-    dtype_and_x,
-    shape,
-    dtype,
-    as_variable,
-    num_positional_args,
-    native_array,
-    frontend,
-    fn_tree,
-    on_device,
-):
-    input_dtype, x = dtype_and_x
-    helpers.test_frontend_function(
-        input_dtypes=input_dtype,
-        as_variable_flags=as_variable,
-        with_out=False,
-        num_positional_args=num_positional_args,
-        native_array_flags=native_array,
-        frontend=frontend,
-        fn_tree=fn_tree,
-        on_device=on_device,
-        a=x[0],
-        dtype=dtype[0],
-        shape=shape,
     )
